requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
moviepy==1.0.3
transformers==4.35.2
torch==2.1.1
diffusers==0.24.0
accelerate==0.24.1
requests==2.31.0
beautifulsoup4==4.12.2
gtts==2.4.0
pydub==0.25.1
numpy==1.24.3
pandas==2.1.3
matplotlib==3.8.2
jinja2==3.1.2
aiofiles==23.2.1
python-dotenv==1.0.0